promote the `nope` message into the eternal grammar
We previously had a special case that requires the server to answer an
invalid `want` message with an empty `have` if no `core` has been
negotiated yet, but it was not defined what the server should do
when the client sends an invalid message that is *not* a `want` message
*before* having negotiated a version of `core`.

Promoting `nope` into the eternal grammar gives the server a consistent
method of replying to any invalid messages regardless of the
negotiation state.

To allow for future extension, add language to the specification of
`nope` that allows any module to override it with a more specific error
reporting mechanism if server and client agree to that module.