core1.0: refactor and refine specification of stdio and multiplexed mode
As discussed in #32 (the PR for the vt6/term and vt6/mono modules), this
moves the spawning of stdios into core, with the new core.to-stdio
message.

While writing the spec for the core.to-stdio message, it occurred to me
that the dichotomy between "normal-mode clients" and "multiplexed-mode
clients" is misplaced. A normal-mode client (such as the SSH client)
must be able to spawn a bidirectional byte stream that is then used as
the stdio of a multiplexed-mode program. This commit moves the modes to
the level of individual bidirectional byte streams to fix this.

A bidirectional byte stream usually starts out in message mode, can be
moved into stdio mode with the core.to-stdio message, and upgraded to
multiplexed mode when the client cannot spawn additional server connections.

The upgrade procedure for entering multiplexed mode changes as follows:

- The initial `<ESC>[6V` need not be the first thing that the client
  writes. Consider the case of SSH: The SSH server may print a login
  message (e.g. /etc/motd) or other log messages before transferring
  control to the shell (or the muxer that provides the VT6 server on the
  remote side), so we should allow for `<ESC>[6V` to happen after text
  has already been printed to stdout.

- Since this now means that stdio may be used for output before the
  upgrade to multiplexed mode, we must assume that it is also used for
  input. To avoid ambiguity or race conditions, we need to mark the
  point where the server has seen the `<ESC>[6V` and stdin switches to
  multiplexed mode. The server therefore now has to reply with
  `<ESC>[6V` if it agrees to the upgrade.

- A problem with this approach is that the client may now be blocked
  if the terminal is not VT6-capable and just ignores the `<ESC>[6V`.
  To somewhat ameloriate this issue, there is a new magic string,
  `<ESC>[0V`, that legacy terminals can use to explicitly reject the
  upgrade to multiplexed mode. A "SHOULD" is added with advise how SSH
  (or other transport-level programs) can improve the user experience
  when remote VT6 clients connect to legacy terminals.

Since the term "bidirectional byte stream" now becomes much more
prominent, the term "server connection" is redefined as "bidirectional
byte stream in message mode".